Add fill phone book, add test it.
