+++ src/Lesson3/Main.java
@@ -11,6 +11,14 @@ public class Main {
         UniqWords.printUniqWord(words);
         UniqWords.printCountWord(words);
 
+        PhoneBook ph = new PhoneBook();
+        ph.add(123,"Ivanov");
+        ph.add(124,"Petrov");
+        ph.add(125,"Sidorov");
+        ph.add(126,"Sidorov");
+        ph.add(127,"Smirnoff");
+        ph.add(128,"Smirnoff");
 
+        ph.get("Smirnoff");
     }
 }
